backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlmodel
sqlalchemy
pyjwt
